ticket-sys1/ticketing-system/requirements.txt
flask
mysql-connector-python
redis
